requirements.txt
pycryptodome
Flask
SQLAlchemy
Flask-SQLAlchemy
Werkzeug